rename common/ module to shared/ (no behavior change)
