+++ src/shared/communication_type.rs
