+++ src/shared/diagnostic_session_type.rs
