+++ src/shared/dtc_snapshot.rs
