+++ src/shared/primitive_generics.rs
