+++ src/shared/reset_type.rs
